poss__pager/pos-backend/requirements.txt
fastapi==0.133.1
uvicorn==0.41.0
sqlalchemy==2.0.47
aiomysql==0.3.2
pymysql==1.1.2
python-dotenv==1.2.1
python-jose==3.5.0
passlib==1.7.4
bcrypt==4.0.1
pydantic==2.12.5
pydantic-settings==2.13.1
python-multipart==0.0.20
email-validator==2.3.0
pyserial==3.5
cryptography==46.0.5  # Add this
